Sequence Animation added; include directives changed; shiltoad of minor code refactorings; bvApp has two sets of predefined modes now; fixed some memory leaks; added some test code; reimplemented timer; added basic timer stats with moving average; changed frame timing method a little; changed interfaces a little to suit animation needs; added query instead of dynamic cast; minor changes in samplers; reimplemented default texture mapping via convex hull in SimpleAnimationPlugin and possibly some more but not so relevant

+++ app/xml/pomocnicze/auxilary.cpp
@@ -1,17 +1,24 @@
 #include "auxilary.hpp"
 
-std::string get_file_contents(const char *filename)
+// *********************************
+//
+std::string get_file_contents( const char *filename )
 {
-  std::ifstream in(filename, std::ios::in | std::ios::binary);
-  if (in)
-  {
-    std::string contents;
-    in.seekg(0, std::ios::end);
-    contents.resize(in.tellg());
-    in.seekg(0, std::ios::beg);
-    in.read(&contents[0], contents.size());
-    in.close();
-    return(contents);
-  }
-  throw(errno);
+    std::ifstream in( filename, std::ios::in | std::ios::binary );
+
+    if( in )
+    {
+        std::string contents;
+
+        in.seekg( 0, std::ios::end );
+        contents.resize( (unsigned int) in.tellg() );
+        in.seekg( 0, std::ios::beg );
+        in.read( &contents[0], contents.size() );
+        in.close();
+
+        return(contents);
+    }
+
+    throw( errno );
+
 }
